AP-6411: Further refactoring of the JS Scorer view
The JS code now uses the knime table JS library.
The rates for each row and column of the confusion matrix are now computed in the JS code.
The class statistics html table is created from the JSON datatable.
The overall statistics html table is created from the JSON datatable.
The number of rows, class names and column/row totals as rates are now all inferred from the
3 JSON datables (ConfusionMatrix, ClassStatitistics and OverallStatistics).

AP-6411 (JS Scorer Node)