AP-13019: Add input filtering to Column Selection Configuration
This commit also fixes a bug in the Column Selection Configuration which
always auto-configures if the column chosen by the user is no longer
missing. Instead the node should prompt the user to select a new column.
Note that this fix requires to deprecate the node.

AP-13019 (Input Filtering in Column Selection and Column Filter
Configuration)